+++ src/tokens/ios/MuiBrandTokens.swift
@@ -2,7 +2,7 @@
 // MuiBrandTokens.swift
 //
 
-// Don't edit directly • Generated on Mon, 29 Jun 2026 08:03:41 GMT • muibook.com
+// Don't edit directly • Generated on Mon, 29 Jun 2026 15:21:17 GMT • muibook.com
 
 import UIKit
 
@@ -65,8 +65,8 @@ public class BrandTokens {
     public static let whiteOpacity_80 = UIColor(red: 1.000, green: 1.000, blue: 1.000, alpha: 0.800)
     public static let whiteOpacity_90 = UIColor(red: 1.000, green: 1.000, blue: 1.000, alpha: 0.902)
     public static let whiteOpacity_100 = UIColor(red: 1.000, green: 1.000, blue: 1.000, alpha: 1.0)
-    public static let grey_100 = UIColor(red: 0.961, green: 0.961, blue: 0.961, alpha: 1.0)
-    public static let grey_150 = UIColor(red: 0.922, green: 0.922, blue: 0.922, alpha: 1.0)
+    public static let grey_50 = UIColor(red: 0.961, green: 0.961, blue: 0.961, alpha: 1.0)
+    public static let grey_100 = UIColor(red: 0.922, green: 0.922, blue: 0.922, alpha: 1.0)
     public static let grey_200 = UIColor(red: 0.898, green: 0.898, blue: 0.898, alpha: 1.0)
     public static let grey_300 = UIColor(red: 0.800, green: 0.800, blue: 0.800, alpha: 1.0)
     public static let grey_400 = UIColor(red: 0.667, green: 0.667, blue: 0.667, alpha: 1.0)
@@ -75,9 +75,11 @@ public class BrandTokens {
     public static let grey_700 = UIColor(red: 0.333, green: 0.333, blue: 0.333, alpha: 1.0)
     public static let grey_800 = UIColor(red: 0.267, green: 0.267, blue: 0.267, alpha: 1.0)
     public static let grey_900 = UIColor(red: 0.200, green: 0.200, blue: 0.200, alpha: 1.0)
-    public static let grey_1000 = UIColor(red: 0.133, green: 0.133, blue: 0.133, alpha: 1.0)
-    public static let grey_1100 = UIColor(red: 0.067, green: 0.067, blue: 0.067, alpha: 1.0)
-    public static let grey_1200 = UIColor(red: 0.020, green: 0.020, blue: 0.020, alpha: 1.0)
+    public static let grey_1000 = UIColor(red: 0.165, green: 0.165, blue: 0.165, alpha: 1.0)
+    public static let grey_1100 = UIColor(red: 0.133, green: 0.133, blue: 0.133, alpha: 1.0)
+    public static let grey_1200 = UIColor(red: 0.098, green: 0.098, blue: 0.098, alpha: 1.0)
+    public static let grey_1300 = UIColor(red: 0.067, green: 0.067, blue: 0.067, alpha: 1.0)
+    public static let grey_1400 = UIColor(red: 0.020, green: 0.020, blue: 0.020, alpha: 1.0)
     public static let red_100 = UIColor(red: 1.000, green: 0.945, blue: 0.933, alpha: 1.0)
     public static let red_200 = UIColor(red: 1.000, green: 0.835, blue: 0.804, alpha: 1.0)
     public static let red_300 = UIColor(red: 1.000, green: 0.620, blue: 0.588, alpha: 1.0)